Better readme, fixed inactivate
I wasn't unpacking enough values, because the "did bug" message doesn't get joined back together, meaning the contents of the msg_args lists are different for the two cases of new/inactivate. MAYBE I'll make it more consistent in the future -- maybe that is the hobgoblin of a little mind.

Signed-off-by: Alexander Loewi <amloewi@gmail.com>