Exception handling on "send_message"
Also, added some more bug names, and randomized name assignment. For fun. Should still be able to handle an arbitrary number of them.

The error handling comes from an error that crashed everything, with the message "HTTP 400 error: 21610: The message From/To pair violates a blacklist rule." Didn't know what to do -- just wrapped it all in an exception that I print when it balks. Sorry, blacklisted people -- I can't even message you to say you're on the blacklist!

Signed-off-by: Alexander Loewi <amloewi@gmail.com>